refactor(unique-mcp): move merge_tool_meta to meta/part alongside MetaPart
