+++ unique_mcp/src/unique_mcp/meta/part.py
@@ -10,4 +10,24 @@ class MetaPart(Protocol):
     def merge_into_meta(self, meta: dict[str, Any]) -> None: ...
 
 
-__all__ = ["MetaPart"]
+def merge_tool_meta(
+    base: dict[str, Any] | None,
+    *parts: MetaPart,
+) -> dict[str, Any]:
+    """Merge base meta with zero or more MetaPart contributions.
+
+    Args:
+        base: Existing meta dict (e.g. ``{"unique.app/icon": "search"}``).
+        *parts: Any number of :class:`MetaPart` instances; each contributes
+            one key to the merged dict.
+
+    Returns:
+        New dict — ``base`` is never mutated.
+    """
+    out: dict[str, Any] = dict(base or {})
+    for part in parts:
+        part.merge_into_meta(out)
+    return out
+
+
+__all__ = ["MetaPart", "merge_tool_meta"]
